Move Composable back into LoyaltyCardImageViewActivity
compose currently only houses the reusable components, not the "main
UI". For consistency, it is better after all to only put reusable
components there.
